# Hestia-Swagger/swagger.yml
openapi: 3.0.3
info:
  title: Hestia — API Reference
  description: |
    Aggregated API reference for all Hestia services.
    Each path overrides the server URL to route to the correct service.

    | Service   | Port  |
    |-----------|-------|
    | Hub       | 19001 |
    | Archive   | 19002 |
    | Hecate    | 19003 |
    | Oracle    | 19004 |
    | Hermes    | 19005 |
    | Scout     | 19006 |
    | Chronos   | 19007 |
    | Argus     | 19008 |
    | Iris      | 19012 |
    | Hephaestus| 19010 |
    | Dummy     | 19011 |
  version: 1.0.0
servers:
  - url: http://localhost:19001
    description: Hub (default)
...
components:
  schemas:
...
    SubscriptionRequest:
      type: object
      required: [subscription_id, owner, domain, event_type, filters, channels]
      properties:
        subscription_id:
          type: string
        owner:
          type: string
        domain:
          type: string
        event_type:
          type: string
        filters:
          type: object
          additionalProperties: true
        channels:
          type: array
          items:
            type: string
        is_active:
          type: boolean
          default: true
...
paths:
...
  /api/logs:
    get:
      tags: [Hub]
      summary: Retrieve buffered runtime logs (service-local endpoint)
      description: |
        Implemented in all runtime services.
        Current instances:
        - Hub (19001)
        - Archive (19002)
        - Hecate (19003)
        - Oracle (19004)
        - Hermes (19005)
        - Scout (19006)
        - Chronos (19007)
        - Argus (19008)
        - Athena (19009)
        - Hephaestus (19010)
        - Telegram control API (internal: hestia_telegram:8010)

        Sensitive values (tokens, passwords, secrets, credential URLs, auth headers)
        are redacted before entries are buffered and returned.

        Use query filters to keep output dense and useful for troubleshooting.
      servers:
        - url: http://localhost:19001
          description: Hub
        - url: http://localhost:19002
          description: Archive
        - url: http://localhost:19003
          description: Hecate
        - url: http://localhost:19012
          description: Iris
        - url: http://localhost:19004
          description: Oracle
        - url: http://localhost:19005
          description: Hermes
        - url: http://localhost:19006
          description: Scout
        - url: http://localhost:19007
          description: Chronos
        - url: http://localhost:19008
          description: Argus
        - url: http://localhost:19009
          description: Athena
        - url: http://localhost:19010
          description: Hephaestus
        - url: http://hestia_telegram:8010
          description: Telegram control API (Docker network)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            maximum: 2000
        - name: level
          in: query
          schema:
            type: string
            example: INFO
        - name: contains
          in: query
          schema:
            type: string
            example: outbound_event_id=p2-4-smoke-3
      responses:
        "200":
          description: "{service, count, logs[]}"
...
  /api/memory/active:
    get:
      tags: [Archive — Memory]
      summary: List active preferences
      servers:
        - url: http://localhost:19002
      parameters:
        - name: domain
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of preferences
...
  /api/subscriptions:
    post:
      tags: [Archive — Subscriptions]
      summary: Create a notification subscription
      servers:
        - url: http://localhost:19002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequest"
      responses:
        "200":
          description: Subscription
  /api/subscriptions/active:
    get:
      tags: [Archive — Subscriptions]
      summary: List active subscriptions
      servers:
        - url: http://localhost:19002
      parameters:
        - name: domain
          in: query
          schema:
            type: string
        - name: event_type
          in: query
          schema:
            type: string
        - name: owner
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of subscriptions
...
  /api/dispatch/logs/enriched:
    get:
      tags: [Archive — Dispatch Logs]
      summary: List dispatch logs enriched with entity metadata
      servers:
        - url: http://localhost:19002
      parameters:
        - name: subscription_id
          in: query
          schema:
            type: string
        - name: entity_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 15
        - name: hours
          in: query
          schema:
            type: integer
            default: 72
      responses:
        "200":
          description: Enriched log entries
...
  /api/feedback:
    post:
      tags: [Oracle]
      summary: Persist explicit quality feedback for a response/session
      servers:
        - url: http://localhost:19004
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                interaction_id:
                  type: string
                source_client:
                  type: string
                quality_label:
                  type: string
                  enum: [excellent, good, mixed, poor, rejected]
                quality_score:
                  type: integer
                  minimum: 1
                  maximum: 5
                outcome_label:
                  type: string
                feedback_text:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                payload:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: "{ok, feedback}"
    get:
      tags: [Oracle]
      summary: List Oracle feedback records
      servers:
        - url: http://localhost:19004
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
        - name: quality_label
          in: query
          schema:
            type: string
            enum: [excellent, good, mixed, poor, rejected]
        - name: source_client
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: "{count, feedback}"
...
  /api/module/maintenance/reconcile:
    post:
      tags: [Maintenance]
      summary: Execute module maintenance reconcile using the standard contract
      servers:
        - url: http://localhost:19002
        - url: http://localhost:19006
        - url: http://localhost:19007
        - url: http://localhost:19011
        - url: http://localhost:19012
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                task_id:
                  type: string
                  nullable: true
                issue:
                  type: string
                  nullable: true
                requested_action:
                  type: string
                  nullable: true
                environment:
                  type: string
                  default: dev
                dry_run:
                  type: boolean
                  default: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: "{status, service, dry_run, task_id, executed_at, retriable, summary, mutation_count, details}"
...